requirements.txt
aiogram==3.13.1
pydantic==2.9.2
python-dotenv==1.0.1
uvloop==0.20.0 ; sys_platform != "win32"
